refactor: modify pages and components file namings
