+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { createBrowserRouter, RouterProvider } from 'react-router'
-import ChatPage from './components/conversation/pages/chat-page'
-import DefaultPage from './components/conversation/pages/default-page'
+import ChatPage from './components/conversation/pages/ChatPage'
+import DefaultPage from './components/conversation/pages/DefaultPage'
 import RootLayout from './pages/Root'
 
 const router = createBrowserRouter([
