+++ src/pages/Root.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Conversation from '../components/conversation/conversation'
-import Sidebar from '../components/sidebar/sidebar'
+import Conversation from '../components/conversation/Conversation'
+import Sidebar from '../components/sidebar/Sidebar'
 
 export default function RootLayout() {
     return (
